% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Declination-Altitude-functions.R
\name{AppAltfromTopoAlt}
\alias{AppAltfromTopoAlt}
\title{Determine Apparent altitude from Topocentric altitude}
\usage{
AppAltfromTopoAlt(TopoAlt, TempE = TempDefault,
  PresE = PressureDefault)
}
\arguments{
\item{TopoAlt}{the Topocentric altitude (deg), double}

\item{TempE}{Temperature at height eyes (C, default=15C), double}

\item{PresE}{Airpressure at height eyes (mbar, default=1013.25mbar), double}
}
\value{
AppAlt de Apparent altitude (deg), double
}
\description{
The astronomical refraction is determined using the inverse of Sinclair's formula. This formula is based
on the International Standard Atmosphere (ISA: which by the way does not mean the 'average' atmosphere;-).
Due to the uncertainties aorund the atmsopheric condition of the Boundary Layers, the calcuated
refractions will not be very accurate (say with 0.5deg).
Below a Topocentric altitude of -3.5deg, the Apparent altitude is made equal.
}
\examples{
AppAltfromTopoAlt(0,15,1013.25)
#0.4721438916

AppAltfromTopoAlt(c(-4,-0.5,0,2,20))
#-4.00000000000  0.04956607457  0.47214389159  2.27870879488 20.04373829337

}
\references{
Bennett, G.G. 1982. 'The calculation of astronomical refraction in marine navigation',
Journal of Inst. navigation, Vol 35: pp. 255-59.
}
\seealso{
\url{http://www.archaeocosmology.org/eng/refract.htm}
}
\author{
Victor Reijs, \email{lists@archaeocosmology.org}
}
