% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Declination-Altitude-functions.R
\name{TopoAltfromAppAlt}
\alias{TopoAltfromAppAlt}
\title{Determine topocentric Aaltiude from apparent altitude}
\usage{
TopoAltfromAppAlt(AppAlt, TempE = TempDefault, PresE = PressureDefault)
}
\arguments{
\item{AppAlt}{the apparent altitude (deg), double}

\item{TempE}{Temperature at Height eyes (C), double}

\item{PresE}{Airpressure at height eye (mbar), double}
}
\value{
TopoAlt, de topocentric altitude (deg), double
}
\description{
blabla
}
\examples{
TopoAltfromAppAlt(0,15,1013.25)

}
